@@ -1,17 +1,23 @@
 {
   "name": "kinesis-wallet",
   "productName": "Kinesis Wallet",
+  "repository": "https://github.com/KinesisNetwork/wallet-desktop",
   "version": "0.1.0",
   "description": "Stellar compliant hard wallet",
   "main": "src/index.ts",
   "scripts": {
-    "sass-compile": "node sass-compile.js",
+    "pretest": "npm run sass-compile",
+    "test": "npm run lint && npm run compile:tsc && npm run package",
+    "predeploy": "npm run sass-compile",
+    "deploy": "electron-forge publish",
+    "sass-compile": "node-sass --include-path=node_modules/bulma src/style.scss src/style.css",
     "prestart": "npm run sass-compile",
-    "prepublish": "npm run sass-compile",
     "start": "electron-forge start",
     "package": "electron-forge package",
     "make": "electron-forge make",
-    "lint": "tslint --project tsconfig.json --force"
+    "compile:tsc": "tsc",
+    "watch:tsc": "npm run compile:tsc -- -w",
+    "lint": "tslint --project tsconfig.json"
   },
   "keywords": [],
   "author": "ABX",
@@ -43,7 +49,8 @@
         ]
       },
       "electronPackagerConfig": {
-        "packageManager": "npm"
+        "packageManager": "npm",
+        "ignore": "src\\style\\.scss"
       },
       "electronWinstallerConfig": {
         "name": "KinesisWallet"
@@ -51,7 +58,7 @@
       "electronInstallerDebian": {},
       "electronInstallerRedhat": {},
       "github_repository": {
-        "owner": "bullioncapital",
+        "owner": "kinesisnetwork",
         "name": "kinesis-wallet"
       },
       "s3": {
@@ -61,25 +68,35 @@
     }
   },
   "dependencies": {
-    "@types/electron-devtools-installer": "^2.0.2",
-    "bulma": "^0.6.2",
+    "bulma": "^0.7.1",
     "crypto-js": "^3.1.9-1",
     "electron-compile": "^6.4.2",
     "electron-devtools-installer": "^2.1.0",
     "electron-json-storage": "^4.0.2",
     "electron-squirrel-startup": "^1.0.0",
-    "js-kinesis-sdk": "^0.9.1",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "js-kinesis-sdk": "^0.9.2",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
     "react-hot-loader": "^3.0.0-beta.6",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0",
+    "redux-observable": "^1.0.0-alpha.2",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "sweetalert": "^2.1.0",
-    "tslib": "^1.9.0"
+    "tsconfig-paths": "^3.3.1",
+    "tslib": "^1.9.0",
+    "typesafe-actions": "^2.0.0-1"
   },
   "devDependencies": {
+    "@types/crypto-js": "^3.1.39",
     "@types/electron": "^1.6.10",
+    "@types/electron-devtools-installer": "^2.0.3",
     "@types/electron-json-storage": "^4.0.0",
-    "@types/react": "^15.6.14",
-    "@types/react-dom": "^15.5.7",
+    "@types/lodash": "^4.14.108",
+    "@types/react": "^16.3.13",
+    "@types/react-dom": "^16.0.5",
+    "@types/react-redux": "^6.0.0",
     "babel-plugin-transform-async-to-generator": "^6.24.1",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
@@ -89,6 +106,7 @@
     "electron-prebuilt-compile": "1.7.11",
     "node-sass": "^4.7.2",
     "tslint": "^5.1.0",
-    "typescript": "^2.7.2"
+    "tslint-react": "^3.6.0",
+    "typescript": "^2.8.3"
   }
 }
