@@ -1,35 +1,45 @@
-import { app, BrowserWindow } from 'electron';
-import installExtension, { REACT_DEVELOPER_TOOLS } from 'electron-devtools-installer';
-import { enableLiveReload } from 'electron-compile';
-const path = require('path')
+import { app, BrowserWindow } from 'electron'
+import { enableLiveReload } from 'electron-compile'
+import installExtension, { REACT_DEVELOPER_TOOLS, REDUX_DEVTOOLS } from 'electron-devtools-installer'
+import * as path from 'path'
+import { register } from 'tsconfig-paths'
+
+// tslint:disable-next-line:no-var-requires no-require-imports
+const config = require('../tsconfig.json')
+
+register({
+  baseUrl: config.compilerOptions.baseUrl,
+  paths: config.compilerOptions.paths,
+})
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
-let mainWindow: Electron.BrowserWindow | null = null;
+let mainWindow: Electron.BrowserWindow | null = null
 
-const isDevMode = process.execPath.match(/[\\/]electron/);
+const isDevMode = process.execPath.match(/[\\/]electron/)
 
 if (isDevMode) {
-  enableLiveReload({strategy: 'react-hmr'});
+  enableLiveReload({strategy: 'react-hmr'})
 }
 
 const createWindow = async () => {
   // Create the browser window.
   mainWindow = new BrowserWindow({
-    minHeight: 800,
-    minWidth: 1200,
-    height: 800,
-    width: 1200,
-    icon: path.join(__dirname, 'icons/png/64x64.png')
-  });
+    height: 768,
+    icon: path.join(__dirname, 'icons/png/64x64.png'),
+    minHeight: 768,
+    minWidth: 1024,
+    width: 1024,
+  })
 
   // and load the index.html of the app.
-  mainWindow.loadURL(`file://${__dirname}/index.html`);
+  mainWindow.loadURL(`file://${__dirname}/index.html`)
 
   // Open the DevTools.
   if (isDevMode) {
-    await installExtension(REACT_DEVELOPER_TOOLS);
-    mainWindow.webContents.openDevTools();
+    await installExtension(REACT_DEVELOPER_TOOLS)
+    await installExtension(REDUX_DEVTOOLS)
+    mainWindow.webContents.openDevTools()
   }
 
   // Emitted when the window is closed.
@@ -37,31 +47,31 @@ const createWindow = async () => {
     // Dereference the window object, usually you would store windows
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
-    mainWindow = null;
-  });
-};
+    mainWindow = null
+  })
+}
 
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-app.on('ready', createWindow);
+app.on('ready', createWindow)
 
 // Quit when all windows are closed.
 app.on('window-all-closed', () => {
   // On OS X it is common for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
   if (process.platform !== 'darwin') {
-    app.quit();
+    app.quit()
   }
-});
+})
 
 app.on('activate', () => {
   // On OS X it's common to re-create a window in the app when the
   // dock icon is clicked and there are no other windows open.
   if (mainWindow === null) {
-    createWindow();
+    createWindow()
   }
-});
+})
 
 // In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and import them here.
